users: structure, model, dtos, exception, controller, mapper,service, tests

+++ main-service/src/main/java/ru/practicum/service/UserServiceImpl.java
@@ -1,11 +1,13 @@
 package ru.practicum.service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.dto.NewUserRequestDto;
 import ru.practicum.dto.UserDto;
+import ru.practicum.exception.NotFoundException;
 import ru.practicum.mapper.UserMapper;
 import ru.practicum.model.User;
 import ru.practicum.repository.UserRepository;
@@ -15,6 +17,7 @@ import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
+@Slf4j
 @Transactional(readOnly = true)
 
 public class UserServiceImpl implements UserService {
@@ -24,7 +27,7 @@ public class UserServiceImpl implements UserService {
     @Override
     public List<UserDto> getUsers(List<Long> ids, Pageable pageable) {
         if (ids == null || ids.isEmpty()) {
-            return userRepository.findAll()
+            return userRepository.findAll(pageable)
                     .stream()
                     .map(userMapper::mapUserToUserDto)
                     .collect(Collectors.toList());
@@ -37,14 +40,19 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public NewUserRequestDto addUser(NewUserRequestDto newUserRequestDto) {
+    public UserDto addUser(NewUserRequestDto newUserRequestDto) {
         User user = userMapper.mapNewUserRequestDtoToUser(newUserRequestDto);
-        return userMapper.mapUserToNewUserRequestDto(userRepository.save(user));
+        return userMapper.mapUserToUserDto(userRepository.save(user));
     }
 
     @Override
     @Transactional
     public void delete(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            log.info("User with id: {} was not found", userId);
+            throw new NotFoundException("User not found");
+        }
         userRepository.deleteById(userId);
+        log.info("User with id: {} was deleted", userId);
     }
 }
